+++ docs/src/components/ComponentDoc/PerformanceSection.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react'
+import ExampleSection, { ExampleSectionProps } from './ExampleSection'
+
+const PerformanceSection: React.FC<ExampleSectionProps> = props => {
+  return process.env.NODE_ENV === 'development' ? <ExampleSection {...props} /> : null
+}
+
+export default PerformanceSection