+++ docs/src/examples/components/Dropdown/index.tsx
@@ -1,11 +1,16 @@
 import * as React from 'react'
+
 import Types from './Types'
 import Variations from './Variations'
+import State from './State'
+import Usage from './Usage'
 
 const DropdownExamples = () => (
   <div>
     <Types />
     <Variations />
+    <State />
+    <Usage />
   </div>
 )