+++ docs/src/examples/components/Indicator/Types/IndicatorExampleDirection.shorthand.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react'
+import { Indicator } from '@stardust-ui/react'
+
+const IndicatorExampleDirection = () => (
+  <>
+    <Indicator direction="end" /> <Indicator direction="bottom" /> <Indicator direction="start" />{' '}
+    <Indicator direction="top" />{' '}
+  </>
+)
+
+export default IndicatorExampleDirection