+++ src/components/Grid/Grid.tsx
@@ -9,6 +9,7 @@ import {
   ChildrenComponentProps,
   commonPropTypes,
   ContentComponentProps,
+  rtlTextContainer,
 } from '../../lib'
 import { ReactProps } from '../../../types/utils'
 import { Accessibility } from '../../lib/accessibility/types'
@@ -72,7 +73,11 @@ class Grid extends UIComponent<ReactProps<GridProps>, any> {
     const { children, content } = this.props
 
     return (
-      <ElementType className={classes.root} {...unhandledProps}>
+      <ElementType
+        className={classes.root}
+        {...rtlTextContainer.getAttributes({ forElements: [children, content] })}
+        {...unhandledProps}
+      >
         {childrenExist(children) ? children : content}
       </ElementType>
     )