+++ src/components/List/ListItem.tsx
@@ -7,6 +7,7 @@ import {
   UIComponentProps,
   commonPropTypes,
   ContentComponentProps,
+  isFromKeyboard,
 } from '../../lib'
 import ItemLayout from '../ItemLayout/ItemLayout'
 import { listItemBehavior } from '../../lib/accessibility'
@@ -44,12 +45,23 @@ export interface ListItemProps extends UIComponentProps, ContentComponentProps<a
    * @param {object} data - All props.
    */
   onClick?: ComponentEventHandler<ListItemProps>
+
+  /**
+   * Called after user's focus.
+   * @param {SyntheticEvent} event - React's original SyntheticEvent.
+   * @param {object} data - All props.
+   */
+  onFocus?: ComponentEventHandler<ListItemProps>
+}
+
+export interface ListItemState {
+  isFromKeyboard: boolean
 }
 
 /**
  * A list item contains a single piece of content within a list.
  */
-class ListItem extends UIComponent<ReactProps<ListItemProps>> {
+class ListItem extends UIComponent<ReactProps<ListItemProps>, ListItemState> {
   static create: Function
 
   static displayName = 'ListItem'
@@ -58,7 +70,6 @@ class ListItem extends UIComponent<ReactProps<ListItemProps>> {
 
   static propTypes = {
     ...commonPropTypes.createCommon({
-      children: false,
       content: false,
     }),
     contentMedia: PropTypes.any,
@@ -81,11 +92,16 @@ class ListItem extends UIComponent<ReactProps<ListItemProps>> {
 
     accessibility: PropTypes.func,
     onClick: PropTypes.func,
+    onFocus: PropTypes.func,
   }
 
   static defaultProps = {
     as: 'li',
     accessibility: listItemBehavior as Accessibility,
+  }
+
+  state = {
+    isFromKeyboard: false,
   }
 
   protected actionHandlers: AccessibilityActionHandlers = {
@@ -97,6 +113,11 @@ class ListItem extends UIComponent<ReactProps<ListItemProps>> {
 
   handleClick = e => {
     _.invoke(this.props, 'onClick', e, this.props)
+  }
+
+  handleFocus = (e: React.SyntheticEvent) => {
+    this.setState({ isFromKeyboard: isFromKeyboard() })
+    _.invoke(this.props, 'onFocus', e, this.props)
   }
 
   renderComponent({ classes, accessibility, unhandledProps, styles }) {
@@ -132,6 +153,7 @@ class ListItem extends UIComponent<ReactProps<ListItemProps>> {
         headerMediaCSS={styles.headerMedia}
         contentCSS={styles.content}
         onClick={this.handleClick}
+        onFocus={this.handleFocus}
         {...accessibility.attributes.root}
         {...accessibility.keyHandlers.root}
         {...unhandledProps}