+++ src/components/Menu/MenuItem.tsx
@@ -15,15 +15,17 @@ import {
   commonPropTypes,
   isFromKeyboard,
   EventStack,
+  rtlTextContainer,
 } from '../../lib'
 import Icon from '../Icon/Icon'
 import Menu from '../Menu/Menu'
-import Slot from '../Slot/Slot'
+import Box from '../Box/Box'
 import { menuItemBehavior, submenuBehavior } from '../../lib/accessibility'
 import { Accessibility, AccessibilityActionHandlers } from '../../lib/accessibility/types'
 import { ComponentEventHandler, ReactProps, ShorthandValue } from '../../../types/utils'
 import { focusAsync } from '../../lib/accessibility/FocusZone'
 import Ref from '../Ref/Ref'
+import Indicator from '../Indicator/Indicator'
 
 export interface MenuItemProps
   extends UIComponentProps,
@@ -105,6 +107,9 @@ export interface MenuItemProps
 
   /** Indicates whether the menu item is part of submenu. */
   inSubmenu?: boolean
+
+  /** Shorthand for the submenu indicator. */
+  indicator?: ShorthandValue
 }
 
 export interface MenuItemState {
@@ -144,6 +149,7 @@ class MenuItem extends AutoControlledComponent<ReactProps<MenuItemProps>, MenuIt
     defaultMenuOpen: PropTypes.bool,
     onActiveChanged: PropTypes.func,
     inSubmenu: PropTypes.bool,
+    indicator: customPropTypes.itemShorthand,
   }
 
   static defaultProps = {
@@ -181,8 +187,11 @@ class MenuItem extends AutoControlledComponent<ReactProps<MenuItemProps>, MenuIt
       primary,
       secondary,
       active,
+      vertical,
+      indicator,
       disabled,
     } = this.props
+    const indicatorWithDefaults = indicator === undefined ? {} : indicator
 
     const { menuOpen } = this.state
 
@@ -204,7 +213,14 @@ class MenuItem extends AutoControlledComponent<ReactProps<MenuItemProps>, MenuIt
             Icon.create(this.props.icon, {
               defaultProps: { xSpacing: !!content ? 'after' : 'none' },
             })}
-          {content}
+          {rtlTextContainer.createFor({ element: content })}
+          {menu &&
+            Indicator.create(indicatorWithDefaults, {
+              defaultProps: {
+                direction: vertical ? 'end' : 'bottom',
+                styles: styles.indicator,
+              },
+            })}
         </ElementType>
       </Ref>
     )
@@ -219,13 +235,14 @@ class MenuItem extends AutoControlledComponent<ReactProps<MenuItemProps>, MenuIt
               secondary,
               styles: styles.menu,
               submenu: true,
+              indicator,
             },
           })}
         </Ref>
       ) : null
 
     if (wrapper) {
-      return Slot.create(wrapper, {
+      return Box.create(wrapper, {
         defaultProps: {
           className: cx('ui-menu__item__wrapper', classes.wrapper),
           ...accessibility.attributes.wrapper,