+++ src/components/Popup/Popup.tsx
@@ -175,10 +175,7 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
       this.close(e, () => _.invoke(this.triggerDomElement, 'focus'))
       e.stopPropagation()
     },
-    close: e => {
-      this.close(e)
-      e.stopPropagation()
-    },
+    close: e => this.close(e),
   }
 
   public componentDidMount() {
@@ -234,7 +231,7 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
               this.triggerDomElement && !this.triggerDomElement.contains(e.target)
 
             if (isOutsidePopupElement && isOutsideTriggerElement) {
-              this.state.open && this.trySetOpen(false, e, true)
+              this.state.open && this.trySetOpen(false, e)
             }
           },
           {
@@ -468,10 +465,9 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
     )
   }
 
-  private trySetOpen(newValue: boolean, eventArgs: any, forceChangeEvent: boolean = false) {
-    if (this.trySetState({ open: newValue }) || forceChangeEvent) {
-      _.invoke(this.props, 'onOpenChange', eventArgs, { ...this.props, ...{ open: newValue } })
-    }
+  private trySetOpen(newValue: boolean, eventArgs: any) {
+    this.trySetState({ open: newValue })
+    _.invoke(this.props, 'onOpenChange', eventArgs, { ...this.props, ...{ open: newValue } })
   }
 
   private setPopupOpen(newOpen, e) {
@@ -489,7 +485,7 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
 
   private close = (e, onClose?: Function) => {
     if (this.state.open) {
-      this.trySetOpen(false, e, true)
+      this.trySetOpen(false, e)
       onClose && onClose()
     }
   }